+++ lib/overcommit/hook/pre_commit/hard_tabs.rb
@@ -3,7 +3,7 @@ module Overcommit::Hook::PreCommit
   class HardTabs < Base
     def run
       # Catches hard tabs
-      result = command(%w[grep -IHn] + ["\t"] + applicable_files)
+      result = execute(%w[grep -IHn] + ["\t"] + applicable_files)
       unless result.stdout.empty?
         return :bad, "Hard tabs detected:\n#{result.stdout}"
       end