+++ spec/overcommit/hook/pre_commit/coffee_lint_spec.rb
@@ -22,7 +22,7 @@ describe Overcommit::Hook::PreCommit::CoffeeLint do
     before do
       result = double('result')
       result.stub(:success?).and_return(true)
-      subject.stub(:command).and_return(result)
+      subject.stub(:execute).and_return(result)
     end
 
     it { should pass }
@@ -33,7 +33,7 @@ describe Overcommit::Hook::PreCommit::CoffeeLint do
       result = double('result')
       result.stub(:success?).and_return(false)
       result.stub(:stdout)
-      subject.stub(:command).and_return(result)
+      subject.stub(:execute).and_return(result)
     end
 
     it { should fail_check }